rename caps_fastapi to fastpi in sdk
